--- src/components/feed/AddNode/addnode.scss
@@ -1,170 +1,0 @@
-@import "../../../assets/scss/helpers/variables";
-
-.pf-c-modal-box {
-  width: 50% !important;
-}
-.screen-one {
-  .pf-c-dropdown__toggle-text,
-  .pf-c-dropdown__toggle svg {
-    color: var(--pf-c-dropdown__toggle--Color) !important;
-  }
-  .pf-c-dropdown__toggle {
-    margin-bottom: 0 !important;
-  }
-
-  .pficon .pficon-close {
-    color: #8a8a8a;
-  }
-
-  .pf-c-form__group {
-    margin: 15px 0;
-  }
-
-  .pf-c-form__label {
-    margin-bottom: 10px;
-  }
-
-  .pf-c-form__label-text {
-    color: black;
-    font-size: 15px;
-  }
-}
-
-.plugin-select {
-  color: black !important;
-  border: 1px solid #f1f1f1;
-  box-shadow: none;
-  .pf-c-accordion__toggle,
-  .pf-c-accordion__expanded-content-body {
-    color: black !important;
-  }
-  .pf-c-accordion__toggle {
-    border: none;
-    padding-left: 15px;
-    padding-right: 10px;
-    margin-bottom: 0 !important;
-    .pf-c-accordion__toggle-text {
-      font-size: 16px;
-    }
-    &.pf-m-expanded .pf-c-accordion__toggle-text {
-      font-weight: bold;
-    }
-  }
-
-  #all-content,
-  #recent-content {
-    margin-left: 15px;
-    margin-bottom: 10px;
-  }
-
-  .pf-c-accordion__expanded-content-body {
-    padding-right: 15px !important;
-    padding-left: 15px !important;
-  }
-
-  .plugin-list {
-    max-height: 250px;
-    overflow: auto;
-    padding-right: 10px;
-  }
-
-  .plugin-list-filter {
-    margin-bottom: 5px;
-  }
-
-  .plugin-list li {
-    padding: 10px 5px;
-    cursor: pointer;
-    &:hover {
-      background: #f5f5f5;
-    }
-    &.selected {
-      background: var(--pf-global--primary-color--100);
-      color: white;
-    }
-  }
-
-  .plugin-list-filter {
-    outline: none;
-  }
-}
-
-.modal.modeless-pf.shown {
-  display: block;
-  opacity: 1;
-}
-
-.modal-font-color {
-  color: black;
-}
-
-.modal-footer {
-  // Configure & Add Node buttons
-  button.pf-m-primary {
-    &:hover {
-      background: var(--pf-c-button--m-primary--BackgroundColor) !important;
-      color: white !important;
-    }
-    &::after {
-      content: none !important;
-    }
-  }
-}
-
-.screen-two {
-  .plugin-name {
-    background: white;
-  }
-
-  .param-form {
-    display: flex;
-    flex-direction: row;
-    font-family: monospace !important;
-  }
-
-  .param-form-item-2 {
-    background-color: white !important;
-    color: white !important;
-    font-size: 18px !important;
-    font-family: monospace !important;
-    border: 1px solid black;
-    width: 25% !important;
-  }
-
-  .approved {
-    background-color: green !important;
-  }
-
-  .not-approved {
-    background-color: #a30000 !important;
-  }
-
-  .param-form-item-3 {
-    color: black !important;
-    border: 1px solid gray !important;
-    font-family: monospace !important;
-  }
-  .param-form-item-help {
-    margin-top: -15px !important;
-    background-color: white !important;
-    color: black !important;
-    text-align: left !important;
-  }
-
-  .plugin-input {
-    margin-top: -5px;
-  }
-
-  .errors {
-    font-size: 16px;
-    .error-message {
-      color: #783e02 !important;
-    }
-    svg {
-      color: #ec7909;
-      margin-right: 7px;
-    }
-    margin-top: 10px;
-    margin-bottom: 10px;
-  }
-}
