+++ src/components/feed/AddNode/styles/addnode.scss
@@ -1,4 +1,4 @@
-@import "../../assets/scss/helpers/variables";
+@import "../../../../assets/scss/helpers/variables";
 
 .screen-one {
   .pf-c-dropdown__toggle-text,
